Implemented "equals" button functionality with operator buttons ( / * + - )
added "right-hand-side" value storage and current operator storage.

If the user enters a number followed by an operator, the program awaits the entry of a right-hand-side number.

When the equals (=) button is selected and there is a current operator and a right-hand-value, the calculate() function is called and the total is stored in memory and displayed.

Selecting equals again will use the same operator and right-hand-value that is stored in memory (equals-chaining)

What needs to be added:
Operator chaining, such that the user can enter right-hand-values and select operators in sequence without having to select the equals button to calculate.

The decimal place button is not implemented, so you cannot currently use the number pad buttons to add a decimal point to the input value.